CA-16 - move files around
